Fixes card backface + gamut
