+++ packages/gamut-labs/src/experimental/index.ts
@@ -1,4 +1,3 @@
-export * from './Card';
 export * from './Coachmark';
 export * from './HubCard';
 export * from './Popover';
