+++ packages/gamut/src/DeprecatedCard/CardBody.tsx
