+++ packages/gamut/src/DeprecatedCard/CardContent/index.tsx
