+++ packages/gamut/src/DeprecatedCard/CardContent/styles.module.scss
