+++ packages/gamut/src/DeprecatedCard/CardEyebrow/styles.module.scss
