+++ packages/gamut/src/DeprecatedCard/index.tsx
@@ -1,0 +1,8 @@
+export * from './CardShell';
+export * from './CardBody';
+export * from './CardFooter';
+export * from './CardHeader';
+export * from './CardButton';
+export * from './CardContent';
+export * from './CardEyebrow';
+export * from './CardStack';
