+++ packages/gamut/src/DeprecatedCard/styles/CardBody.module.scss
