+++ packages/gamut/src/DeprecatedCard/styles/CardFooter.module.scss
