+++ packages/gamut/src/index.tsx
@@ -13,9 +13,10 @@ export * from './Button';
 export * from './ButtonDeprecated';
 export * from './ButtonDeprecatedBase';
 export * from './BodyPortal';
-export * from './Card/CardBody';
-export * from './Card/CardFooter';
-export * from './Card/CardShell';
+export * from './Card';
+export * from './DeprecatedCard/CardBody';
+export * from './DeprecatedCard/CardFooter';
+export * from './DeprecatedCard/CardShell';
 export * from './ContentContainer';
 export * from './FlexBox';
 export * from './FlexGrid';
