+++ packages/styleguide/stories/Atoms/Card.stories.mdx
@@ -1,4 +1,4 @@
-import { Card } from '@codecademy/gamut-labs/src';
+import { Card } from '@codecademy/gamut/src';
 import title from '@codecademy/macros/lib/title.macro';
 import { Canvas, Meta, Story } from '@storybook/addon-docs/dist/blocks';
